Tests to demonstrate that NODE KEY behaves as UNIQUENESS
Specifically in the Cypher planner support.
