+++ enterprise/cypher/acceptance-spec-suite/src/test/scala/org/neo4j/internal/cypher/acceptance/CompositeUniquenessConstraintAcceptanceTest.scala
@@ -24,9 +24,7 @@ import org.junit.Assert.assertThat
 import org.neo4j.cypher.javacompat.internal.GraphDatabaseCypherService
 import org.neo4j.cypher._
 import org.neo4j.graphdb.ConstraintViolationException
-import org.neo4j.kernel.GraphDatabaseQueryService
 import org.neo4j.test.TestEnterpriseGraphDatabaseFactory
-import org.scalatest.matchers.{MatchResult, Matcher}
 import org.neo4j.cypher.internal.frontend.v3_2.helpers.StringHelper._
 import org.neo4j.graphdb.config.Setting
 
@@ -80,19 +78,5 @@ class CompositeUniquenessConstraintAcceptanceTest extends ExecutionEngineFunSuit
 
   private def exec(query: String) {
     executeWithCostPlannerAndInterpretedRuntimeOnly(query.fixNewLines).toList
-  }
-
-  case class haveConstraints(expectedConstraints: String*) extends Matcher[GraphDatabaseQueryService] {
-    def apply(graph: GraphDatabaseQueryService): MatchResult = {
-      graph.inTx {
-        val constraintNames = graph.schema().getConstraints.asScala.toList.map(i => s"${i.getConstraintType}:${i.getLabel}(${i.getPropertyKeys.asScala.toList.mkString(",")})")
-        val result = expectedConstraints.forall(i => constraintNames.contains(i.toString))
-        MatchResult(
-          result,
-          s"Expected graph to have constraints ${expectedConstraints.mkString(", ")}, but it was ${constraintNames.mkString(", ")}",
-          s"Expected graph to not have constraints ${expectedConstraints.mkString(", ")}, but it did."
-        )
-      }
-    }
   }
 }
